requirements.txt
torch>=2.0.0
transformers>=4.30.0
polars>=0.20.0
numpy>=1.24.0
scipy>=1.10.0
wfdb>=4.1.0
mne>=1.5.0
ssqueezepy>=0.6.0
Pillow>=10.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0 